refactor(download): move PRIDE-specific helpers into PrideProvider
Tier 1 + Tier 2 of the provider-specific helper audit:

by_url.py (Tier 1) -> PrideProvider:
  - _extract_pride_accession -> PrideProvider.extract_accession_from_url
  - _validate_urls_checksums -> PrideProvider.validate_urls_checksums
  download_files_by_url now calls PrideProvider.validate_urls_checksums when
  checksum_check is set. by_url keeps only generic URL-download mechanics.

util.py (Tier 2) -> PrideProvider:
  - _get_download_url (the aspera/ftp/globus/s3 protocol resolution +
    PRIDE archive HTTPS rewrite for globus) -> PrideProvider._get_download_url
  - _resolve_local_path -> PrideProvider._resolve_local_path
  The base Provider.get_download_url is now generic (returns the
  'FTP Protocol' location value, which is what MassIVE/JPOST/iProX need);
  PrideProvider.get_download_url overrides it with the multi-protocol
  resolution. PRIDE's static download methods call
  PrideProvider._get_download_url directly.

util.py now holds only genuinely generic helpers (Progress, compute_md5,
validate_download, read_checksum_file, _find_tsv_columns, _is_md5_checksum,
_remove_if_exists). Removed the lazy 'from pridepy.download.pride import
PrideProvider' imports that util needed for the moved functions.

Test: the one test referencing util._get_download_url now references
PrideProvider._get_download_url (PrideProvider already imported).

No behaviour change. 68 passed, 4 skipped. flake8 (E9,F63,F7,F82) clean.
Smoke MD5 unchanged.
